fix filtering and separate pagination
